maint: reverted sqrt of lagrange
The lagrange multiplier for a history of 1 is equal
to <R|R> and is thus directly related to the derivatives.

Probably we should do lagrange / <R|R> to ensure it is
correctly bounded. I imagine the lagrange may be
very large for very large systems since <R|R> may also
be very large.

Signed-off-by: Nick Papior <nickpapior@gmail.com>