enh: added grouping of atoms
Sorting now allows grouping of atoms. There are 4 different
groupings, Z, tag, species, symbol.
This is controlled by group=str or group=(str, ...)

where str is one of the "groupings" and ... is a list
or list of lists with items that may be grouped together.
Optionally may one of these items be None in which case
it will be filled with the *rest* of the allowable searchable
items.

For instance

     geom.sort(group=('symbol', None, 'C'))

first groups all non C atoms, then all C atoms.

Signed-off-by: Nick Papior <nickpapior@gmail.com>