enh: added first Dispatcher method
An extension of ideas proposed by Pol.

This continues work on #213 and #214.

This is different than the proposed idea for
a couple of reasons.

1) Both Pol and Jonas mentioned that using asaverage
   was not intuitive since the methods changed the object
   it self AND returned it-self.

   This was a bad decision from me in the first place.

2) I wanted to have some functionality that wasn't there
   yet:
   a) make it clear in the code _when_ a user does dispatching
   b) allow users to customize how dispatching works, e.g. #182
   c) make returned dispatch methods different than the object it self
      I.e. bz.dispatch.average != bz.dispatch.sum

This is all accomplished in this merge request.

It is relatively complicated given what it does.
In sisl/_dispatch.py the basic implementation exists.

There are 3 classes since we need a class instance (basically
a singleton) that holds the registered functions and allows for
registrering more.

When bz.dispatch is called it returns an ObjectDispatcher with
the current allowed methods and the object (bz). At this point it is
pretty much like #214, but by allowing specific keys to retrieve the
dispatch method (or actually class).