maint: changed name to apply in BZ
Changed Yield to Iterator and made the key be apply.iter.
This is more consistent as Jonas correctly mentioned.

Enabled a default dispatcher (for BZ it is iter).

Changed __call__ interface for AbstractDispatcher.
It now returns the function which makes more sense
at then the function may be re-used.

Added an asterisk for the default dispatch class.