maint: enabled Jonas suggestion
Now one can retrieve a MethodDispatcher to allow
methods to have different dispatchers.

This required quite some code in Object/ClassDispatcher
to allow one to correctly obtain the method.
Also because it is difficult to emulate a function
with a class instance (__doc__ cannot easily be overwritten).

Hence the ObjectDispatcher will *always* return a MethodDispatcher
if the dispatch-key does not exist.

This is good and bad since for dispatch-keys which are also attributes
of the object, one will be forced to retrieve the dispatch.
Here one can only bypass this by retrieving the wanting dispatch first
and then the method.