bug: cleaned Hermitian property of FC files
Now I can easily reproduce Vibra eigenvalues
for molecules (bulk to be tested).

This was inspired by Bernhard Kretz who
supplied a test.

Added a H2 molecule test for calculating the dynamical
matrix and eigenmode spectrum of this.

It seems to work by now using the symmetrization
as done in Vibra. They are not 100% the same, but I
think it amounts to the same. Haven't had the time
to fully figure this out... :(