enh: added categorization of geometries
This could be extended for more elaborate uses.
It is currently not that fast but performance could be
improved using the InstanceCache class that wraps
a class and lru_caches methods.

It seems to work rather nice.

One can do complicated things with very little effort.
However, there are things that is not taking into account.
For instance we need to handle "not Category" to do the opposite
category.