bug: fixed oplist changes (for iop), fixed xsf Out argument
The recent b415a8a8c9 commit broke ioperations.
With operators one have to catch the result in case
the value is mutable.

Moved sdata function to sisl/utils/_sisl_cmd.
In the future sisl will be used instead of sdata.
Or should it?

Bug, fixed tbt.Eindex for str retrieval. For strings
it is considered a float and tries to use it as such.
This makes CLI easier since we don't have to check arguments.

Also fixed the latest commit with multiple funcs.