enh: cleaned read_scf from siesta.out
This now successfully reads output and
now also handles cases where the user requests
a minimum number of runs.

Also changed Pol's initial implementation to always
have either imd or iscf as indices in the dataframe.
This is more consistent as it does not rely on data.

Also added a test.