ufunc on SparseCSR and derived types (#226)
* enh: implemented array_ufunc in sparse and sparsegeometry

Now we can use numpy functions on sparse matrices.

There are certain cases where we don't get the correct
data-type. SparseCSR(int) / 2 should be converted to a
float, but currently fails.

This code base properly fails on some other methods but
for now we have a broad range of functionality that is
really nice.

Also I have added a todense method which allows creation
of the full dense matrices (easier to test).