maint: major update on testing framework
xfail's are meant for things that are currently bugs
or for things that are to be implemented (but currently fails for
the same reason).

To check for raises one should do with pytest.raises(...) in
the test.

The above two criteria have now been implemented.
This has reduced xfail's to 2, both fulfill the above xfail
criteria.

This also uncovered some minor bugs (which are corrected).