enh: cleaned ObjectDispatcher
Now one can *call* an ObjectDispatcher to re-create
a new one with different attributes, this allows
for creating parallel dispatchers.

I have also reworked the BZ papply by removing it
and now deferring the apply(pool=True) to the parallel
environment.
This still needs to be documented but I think it works as it
should.