enh: implemented #249
Now Atom(-int) will return an AtomGhost(int) instance.
And Atom(10000) will return an AtomUnknown(10000) instance.

This is because the Ghost procedure is typical only for Siesta
and we need a way to disconnect sisl from Siesta.

Now Atom implements __new__ which decides on which class to produce.

Effectively it does this:

if Z < 0:
   class = AtomGhost
elif Z not in PeriodicTable._Z_int:
   class = AtomUnknown
else:
   class = Atom

The AtomGhost defaults to a mass that is 1e40, if tag is not specified
it will be "ghost".

For AtomUnknown, if tag is not specified
it will be "unknown".

This has the advantage that everything is now handled at the class
inheritance scheme. It is a bit brittle, I think, but easier for
end users. I haven't checked all corner cases, but I think this commit
fixes all issues related.

I have fixed XV, NC and fdf to correctly handle ghosts, there
are probably still some remnants that needs fixing.