maint: removed spin-resolved velocities, merged spin_*_moment and spin_moment
The spin-resolved velocities were nothing but the outer
product of the spin-moment with the velocities. So there
was really no surprise there.

The spin_orbital_moment has been merged in the spin_moment
method with the optional argument project=T|F.
This is easier to maintain and cleaner I think.