enh: enabled sgrid to write out data in tables
Added commands for smooth and interp with more arguments
This forces users to do --interp ... --out because interp
globs all arguments. Slightly annoying, but fine I think.

Also fixed a bug in isosurface which used the wrong transformation
matrix for skewed cells.

Added tests for the smooth arguments and the table-write-out.