maint: changed Z -> zeta for AtomicOrbital, enabled Atom.orbital attr call
Z in AtomicOrbital is changed to zeta. This is to clarify between Atom.Z and
AtomicOrbital.Z. Perhaps somebody was not sure about this?
Currently I have marked .Z as deprecated, but I don't know if it should stay
there.

Also ensured that unknown attributes in Atom gets passed to queries on the
orbitals. This means that .R and .l can be done on Atom.
We still need to do this for Atoms class, but that is a bit more tricky.