enh: initial simplex method addition rough but runs
mnt: cleaning up implementation and greatly amended default.yaml

Now the default.yaml implements references which makes it
very easy to add more elements.

Using references every element only takes up 1 + n,l pairs
number of lines which makes it very readable.
This may be a good starting point for these things.

Enabled reading shell charge confinement.

Ensured that optional values are using defaults from Siesta.
This is important for soft and charge confinement where there
are 1 and 2 optional variables, respectively.

Ensured that when users specify rcore in the AtomPseudo
options it means that core-corrections should be turned on.

Added a plot that plots atom output. Consider moving this
to a separate toolbox to faster plot stuff.

Amended yaml files for configurations to show what can be done.

Also made it clear that core-corrections should be used carelessly.
This is because the total-energy is highly dependent on this due
to the inclusion of core-charge in the simulation.

mnt: cleaned single line

enh: added reading INP from atom to create class

This makes it a bit easier to work with.

Also added plotting utilities to that class
to make it easy to plot stuff.

bug: fixed charge for atom variable and conversion of jac

bug: fixed typo

minor clean-ups

mnt: enabled function runners

enh: enabled the toolbox for plotting atom output

Also added some documentation for this.

mnt: made runner/metric forced keyword arguments

This makes it more versatile, but the siesta minimization more difficult

Needs re-thinking.