mnt: changed all phase signs
All tests still pass with this minor change.
This should accommodate the changes to come for
siesta. I have found that it is necessary to hermitian
conjugate the siesta sparse matrix since they are stored
in CSC and not CSR. This is only a problem for NC/SOC calculations
due to the complex spin-box. So this is why it hasn't
been really necessary before.

Also clarified some documentation here and there.