bug: finally fixed CSC->CSR matrices from Siesta
Basically the phases were incorrect due to transposing the data.
With this, I fix everything and the convention is the same as the
Siesta convention e^ikR (no sign change).
To make this happen I transpose the matrices from Siesta (but not
the spin-box) to ensure I get correct placement of matrix elements.

This also means I can directly compare wavefunctions outputted from
Siesta (wfsxSileSiesta) checks are now consistent.