enh: enabled dataset returns when unzipping data
When using unzip for dataarray we can now return datasets
from xarray. Added a test for this.

Additionally a bug has been fixed for "register" functions
and default values.
I still don't know why this was wrong, but now it works.

BZ.apply.array now also allows unzipping data.
This is necessary when users want to calculate data that
is not commensurate (i.e. different shapes).
Added a test for this.