enh: enabled parsing PAO.Basis block in limited case
The PAO.Basis block can in some cases be parsed correctly
due to the explicit specificiation of ranges.
We allow this, but silently ignore the cases where
we do not know how to do it. I.e. with filtered basis sets.

The ranges are generally also not correct, but it should
be better than not having anything.

This implementation also allows SphericalOrbital to allow
the R argument in case the funcion + grid isn't defined.

Fixes #90.