enh: added dispatchers to Shapes
For conversion between spheres-ellipsoids etc.

This is handy for allowing external shapes to be extended.
Reworked the dispatchers to use a ChainMap to be able to
host the same key depending on the subclasses.

This of course forces the user to think about whether
some previously registered things might not work for
subclasses. But so be it.