tests: removed plotly requirement to test sisl.viz
The current tests only test what is the information that is transferred to the backend,
instead of how is this information plotted. In the future, specific tests may be added
for the backends.

However, backends are not fully untested. If some backend is loadable, the tests will run
for it. These tests do not check if the information is plotted as it should, but at least
the backend will receive a large variety of inputs, which may raise errors if there are
bugs.