enh: greatly speeded up eigen-channel calculation in BTD
Now we use SVD routines to calculate eigenvalues and orthogonal
scattering states for the matrices.
This is many times faster than using the full A for eigenvalue
calculations.
For my test with 76 blocks of size 50 it is >800 times faster.

I have removed the scattering_state_from_* method which is not
really user friendly.