bug: fixed propagate method
Now the propagate method also uses the SVD method to calculate
the scattering states and it works.
This will sometimes be faster since the initial
eigenvalue calculation can quickly limit the search-space
and greatly speed up the calculation contrary to the full
column for the SVD method. However, the cutoff seems more
important here since it is applied twice.