mnt: removed parallel execution
Doing parallel in Python is not easy as it copies
the entire interpreter between processors.

Therefore it is better to have it explicitly set to off
and then let the users decide when to use it.

I have also changed the names to SISL_NUM_PROCS
to match the OpenMP convention.