bug: cleaned the BandStructure class for better control
The algorithm is now very deterministic and limits
double code.

It has changed the arguments a bit.

Plural s are added to clarify when they really
have significance.
Also, the divisions argument is now "divisions between
each point" and so an extra point is added for the end-point.
This may come as a suprise but is stringent.

Also added more tests.