mnt: cleaned and fixed lots of stuff in the reading of WFSX
Generally one should limit the return types as dictionaries
as it imposes naming conventions on certain variables
that may not be optimal for the end user. In these cases
namedtuple returns are better since users can do whatever they
wish with them.

Removed lots of not-used variables in the fortran code.

Changed names of many of the _bin_* methods to explicitly
mention the fortran interface. This clarifies belonging.

Also ran pep8 which fixed lots of indentation problems.