mnt: added fallback version for tarball builds
Added testing script for installing
under various instances.

Now one can download the tarball and still
install since the .gitattributes automatically
stores the git-hash.
This should make it robust for installation on
clusters etc.

Added fallback versions so it always detects something
and should basically never fail.

To run the test script do something like this:

  git archive --output developments/sisl-whatever.tar.gz HEAD
  git archive --output developments/sisl-0.11.0.tar.gz v0.11.0

then they can be tested like this:

  cd developments
  bash test_numpy_installs.sh sisl-whatever.tar.gz

and a couple of tests are runned.