mnt: added simple TB models and made sisl import faster
The dispatch methods were quite heavy since it was processing
the __get__ method all the time. Now the dispatch is stored in
a temporary variable, and the registration on this temp var.

Enabled simple TB models for graphene. The toolbox is not complete
since its use could be extended for more general cases, but for
now this is ok I think.