enh: fully implemented derivative and corrected 2nd order derivatives
The derivatives are now generalized for the state object.
This may or may not be the best approach, however I find it much simpler
to maintain. The sisl.physics.[phonon|electron].velocity and friends
will be deprecated since the derivatives are simpler to calculate from the state
object. Otherwise the lengthy arguments for the 2nd derivative makes them too complex
to actually be used. While it would be nice to keep the functional programming model,
it just does not make sense for this complicated feature.

Now the 2nd derivative should be more correct since we correct it by the:

  2 * ediff * dH/dk

terms.

This also removes the projection method which should simply be an expansion of the norm2
matrix onto each individual state. Therefore it doesn't merit its addition in this method
(for clarity).

This fixes #406.