mnt: cleaned berry-curvature and removed non-used args
The complex argument should not be used, and is thus
completely removed.
Now the conductivity has a new argument `degenerate` which
enables one to decouple degenerate states. This is vital for
the correct calculation of the conductivity.

Also speeded up berry_curvature calculations (and hence
conductivity) by now using matrix multiplications.