enh: enabled SVD calculation of the berry_phase
The berry-phase calculation was previously done for
the simple dot-product. However, in certain cases one
may envision the use of SVD as a more reliable way to
do the dot-products.

Also ensured that the returned values are not forced
into the range ]-pi;pi] by np.angle. Now the user
will have to process these.