enh: enabled slabs to add vacuum at intermediate positions
The algorithm places vacuum (with exact lengths) where the
user denotes, for instance 'ABC ABC ' will place a vacuum
in the middle, and at the end whereas 'ABC ABC' will only place
it in the middle.

This also uncovered a bug in Geometry.add where offset
was not taken into account when a SuperCell was passed.

Added to changelog and docs.