enh: changed cut -> untile, and lots of bug-fixes + perf
The cut method was slow and not accurate in all situations.

Now the method is more robust, and checks for more
corner cases. It warns about incorrect usage and also symmetrizes
when splitting has an even number of supercells. The *final* coupled
region with then be *halved* in contribution since the algorithm
does not know if it will be completely tracked by the left/right side.
This is also what will meen that users should always untile as
much as they can, and then tile to the desired size. This ensures
the best handling of the sparse matrices. One should try and aim
for a zero-warning run.