enh: added multiple replacements for k-points
This is required for replacing large quantities of
k-points. The idea was initiated by @ahkole and polished by @zerothi.

A displacement argument and as_index argument ensures that one
can fine-tune the placement. However, in these cases it is not
strictly checking whether the indices are actually *inside* the
replaced k-points. Probably this should be checked but users
may want to do crazy stuff.

Co-authored-by: A.H. Kole <a.h.kole@uu.nl>
Co-authored-by: Nick Papior <nickpapior@gmail.com>