enh: cleaned mixing methods
The basic functionality and arguments are the same.
The History class is now an attribute in the Mixer
classes that requires them.

The Metric class has been deleted since it didn't really
make sense and a simple func(a, b) argument is required.
This can be made arbitrarily complex anyways.

The history argument can now be made up of an already
made History object such that the same history can be
re-used between different objects. However, it is vital
that users do not call append for each of them requiring
some manual coding around those issues. It will be recommended
to add the argument `append=True` to those __call__ methods
to allow users to use them without appending to the history.

The DIIS mixing method now has an additional argument for the
__call__ to use a different metric vector. Say you want to
calculate the metric using some real vector, but want a complex
vector to be mixed, then one can do this.
The basic principle is this:

  __call__(M, dM, delta)

M is the input matrix
dM is the step of matrix
delta is the metric used for the DIIS mixer

The tests are unchanged but still functional, so I don't think
anything will be broken.