bug: fixed untile for *big* systems
When the orbital connections are greater than 1, part
of the algorithm could fail.

This is now fixed by removing neighbouring duplicates
in the sub_lsc array and finding the largest denominator.
Could be made simpler, but since we don't know about
positive vs. negative findings this will be safer I think.