enh: enabling reading of file-handles and StringIO
This patches the Sile (ascii) formats to enable
reading directly from a TextIOBase, anything else will
be parsed as a filename to be used normally.

This is done using the __new__ method to patch the
class creation and omit the file's opening etc.
Only 2 tests have been added for now, but more should be
added to assert it works on a multitude of things.

Probably the fdf sile will be problematic due to the include
statements that may interfere.

The dynamic class creation will append a "Buffer" suffix
so the class nature is clear. However there will only be some
internal methods which will be overwritten.

This addition was requested and great inspiration from @pfebrer
and @ahkole, thanks!

Also enabled users to overwrite the dynamic creation
of the buffer-class for user-defined classes.
This may come in handy in certain situations where
the buffersile is not enough.