bug: possible bug fix for mixed orbital order
In HSX files the orbitals are created by only knowing the
n, l, zeta values. Hence, the m loop is implicit when
n, l, zeta changes. Previously the check was only if l
changed to some higher value. This commit should be more
stable if one, say, uses the s and d shell (but no p).