mnt: changed index ordering of spin-charge quantities
Before the resulting charge arrays were not strictly the
same order. Now we always have indices such as:

  states, spin, orbitals

which means that spin is not the last index anymore.

Furthermore for polarized calculations we will return
spin dimension as:

  total, Sz

to make it more compatible with the SOC cases.
This will perhaps confuse users in the beginning (those
who were used to the old ordering), but in the long run
this should be easier to handle since users don't need to
put in checks for polarized vs non-colinear calculations.