mnt: finalized the 501 PR
All routines has been reworked and ordered to have the
Cartesian directions as the first dimension.

This might be a bit weird at first, but the benefit is simpler
internal code and a coherence between all routines.

Specifically the PDOS was troublesome since it had varying dimensions
and for polarized it was up/down, whereas now it is total/z.

This will require end-users to accommodate, which is annoying, but needed.

Also fixed bug in displacement calculation.