mnt: major performance boost for COOP curves
Now the sparse matrices are much better utilized.

The new implementation actually uses sparse matrices, whereas
the old expanded the matrices to the full supercell picture, which
meant that a large fraction of the orbital space was never used.

This is now fixed and for regular systems it is easily a factor 100
faster.

For very small systems, it is now much slower, but should be acceptable
since COOP curves are primarily interesting for sizes > 400 orbitals.