mnt: changed oplist to enable generators in loops
This will reduce memory requirements for COP curves
since we do not need to hold all energy-points, only
1 at a time.

This change is now used in COP method and is much easier
on memory (a factor len(E)).

Additionally a tolerance argument to COP has been added
to reduce calculations on energy points where the
contribution is negligeble. This further improves performance
for large systems since there are many states.