feat: added Geometry.apply
A function was necessary to apply operations on segmented
data in such a way to easily change from, say orbital data
to atomic data. This could be either mean/sum/diff calculations.

This is now enabled via the Geometry.apply function.

It will be moved to somewhere more clean in the future.