mnt: ensured mode is passed to read() methods
mode is now a keyword argument in *all* cases.

Also fixed typing for all read methods.
For now these are a bit weird since they are
staticmethods but returns a type "Self". I am not
fully sure how the typing construct infers the self
type when inheritance is done. But perhaps this is
done automatically.
TODO: It might be useful to change all these to
class-methods to allow passing the class type as an
argument to instantiate from a specific class.