mnt: unified the grid reading properties
This ensures that the method by which we reduce
several indices are done in the right manner.

All grid reads should use sisl.io._help.grid_reduce_indices
which can reduce along any axis with a set of factors.
Now the limitation of ensuring the correct length of the
factors has also been lifted. It is only necessary to have
a len(list) <= len(grid.shape[axis]) and the rest will be
understood as 0's.