enh: Added Nodes and Worflows
A new submodule is introduced, sisl.nodes, that implements functionality
to create checkpointable workflows based on nodes.

A notebook to introduce the framework has been added to the documentation.

This framework will be used by the visualization module to make plots
much more modular. It also allows more flexibility and the GUI might be
able to automatically use nodes and workflows (would be great)