bug: fixed swapaxes handling
Now the swapaxes is correct for Cartesian coordinate swaps.

Also extended it to be much more versatile.
One can do collapsed swappings, which might not be that useful.
But rather easy to follow given that I wanted the str arguments.

Thanks to @ahkole for discovering the bug leading to this.

Signed-off-by: Nick Papior <nickpapior@gmail.com>