mnt: changed only= keyword arguments to what=
This is to unify interfaces that has these kinds of arguments.
It affects Geometry|SuperCell.rotate and some tbtrans siles.

To accommodate backwards compatibility for some time I have
added a function to decorate methods enabling a keyword argument
and replacing the *new* keyword argument with the old keyword argument.
I believe users who use the old argument are not aware of the new one
and hence it makes more sense.
A deprecation warning is also issued.

The default for Geometry.rotate is now also changed depending on
whether atoms argument is supplied.
- atoms is None: what="xyz+abc"
- else what="xyz"
when doing sub-rotations it makes more sense to only rotate the
atomic coordinates, while rotations of everything makes more sense
to also rotate the lattice vectors.

Signed-off-by: Nick Papior <nickpapior@gmail.com>