mnt: fixed and simplified some logic in the read_multiple function
It should now be able to handle more cases and more easily transfer
data to a required data-structure.

Added an xyz skip function which is much faster than the read_geometry
call.

Changed default all=True for the ANI file.
Changed the suffix for ANI files to be capitalized.

Added a postprocess function for post-processing the data returned
by the read_multiple function. This would enable one to merge stuff etc.

Signed-off-by: Nick Papior <nickpapior@gmail.com>