bug: reworked argument handling for read_multiple
Also cleaned and fixed xsf to use the new scheme.
This also fixed some bugs in the implementation of the
xsf file.

It should now be more stable and resilient to wrong file
formats.

The start, stop, end arguments are now handled differently
according to all, so all will not overwrite any set values.
This means that all will only change defaults for those not
provided.

Signed-off-by: Nick Papior <nickpapior@gmail.com>