changed SuperCell to Lattice, long overdue
The ambiguity of referencing to the lattice/unit-cell
as a supercell is confusing when we are also dealing with
the description of the *number of supercells*.
Therefore we have changed all SuperCell references and sc
arguments to Lattice and lattice, respectively.

This surely comes at a cost for end-users, since this is
a major transition. I will keep backwards compatibility
with the older implementations for at least a year (perhaps
2). But they should eventually be gone to prefer the lattice.

Signed-off-by: Nick Papior <nickpapior@gmail.com>