moving to scikit-build-core
Everything is now working as intended.
We can build a wheel and install it.
This required quite a bit of cmake codes
to handle installations of the sources.
And scikit-build-core did not have the sufficient
details for f2py handling, so I had to create my own.

Now we have deleted setup.cfg and setup.py

There may still be some hickups, and especially wheel
and sdist content should be revised.

Signed-off-by: Nick Papior <nickpapior@gmail.com>