Finalizing the cmake handling for skip|only in signature creation
Also explicitly added a sparse_io module to offload the write/read
of sparse details.

This is also to make the cmake functionality complete, and it
now works in two ways, either directly specifying details in
the sources, or by adding properties to the source files.
I kind of prefer the latter, but it can be a bit cumbersome.
Let us see.

Signed-off-by: Nick Papior <nickpapior@gmail.com>