cleaned the procedure for locating maxR
Instead of searching for a 0 in the radial function
we now search for an integration that contians 99.99% of the
radial function.
The integration uses the trapezoid integration mechanism
with f(r)**2*r**3 as that should be normalized for a
radial function.

This should make it much more stable against different radial
function arguments and produce more accurate ranges for the
orbitals.

Now a warning will be issued when the routine cannot determine
a correct R (based on inputs).

Ensured that all orbitals with radial functions
can pre-calculate the R in case user did not supply it.
One can always forcefully set it by supplying it.

Signed-off-by: Nick Papior <nickpapior@gmail.com>