preliminary test-bed for slicing io stuff
Fixed documentation issues and enabled this.
Storing sliced objects will also work.
Currently nested slicing is not possible.

enabled nested with of a sile and changed to SileBinder

The sile_read_multiple has been deleted, now the way to
do it is SileBinder.
This is more versatile and allows better handling.
It also patches the documentation.

This changes the behaviour of all files using this.
By default a read_* with the SileBinder will
only read the next entry, however to read
multiple one should use read_*[...] for slicing
the data.
The bound method will try and be smart about only reading
the needed elements, there is a bit of overhead, but
largely it should be fast.

All tests are functional again.

Signed-off-by: Nick Papior <nickpapior@gmail.com>