fixed lattice handling in Grid
Previously one could tile non-commensurate grids.
A commensurate grid has a lattice and a geometry that
are integer repetitions of the original geometry.
This is now checked before one can tile the grid.

The Grid.__str__ has been updated to highlight this
configuration where the commensurability is shown
(instead of lattice). However, when they are not
commensurate, the full lattice + full geometry is shown.

Removed some set_lattice when calling this on LatticeChild.
It had the side-effect of setting the lattice for some
hamiltonians etc when calling wavefunction.

Signed-off-by: Nick Papior <nickpapior@gmail.com>