ensured that wavefunctions do not translate coordinates
Now if a translation would have shifted the atoms,
an information notice will be shown.

Users should in this case try and shift everything into
the primary unit cell which should disable the warning
and ensure proper function.

Ideally this should be solved otherwise, and will
probably be solved later, but for now this seems
like the most appropriate thing to do.

Fixes #592.

Signed-off-by: Nick Papior <nickpapior@gmail.com>