enh: Refactoring the viz module to a more modular design
Everything is now based on functions that adhere to functional programming.
We combine these functions to create workflows and then use the sisl.nodes
framework to provide the update functionality.
Plotting backends are now implemented each on a single `Figure` class that
implements generic drawing methods, as well as subplots, animations... functionality.