fixed transition of the dispatch class
Now the dispatch class is interactive in the sense
that it only adds the fields you are interested in.

class A(_Dispatchs, dispatchs=("hello", "new"))

will add the A.hello and A.new fields.
This makes it alot easier to manage.

The subclassing of a dispatcher class with fields
will now search through the bases and discover the dispatch
names assigned at prior bases to collect them.
Then the dispatch attributes will be handled through the
when_subclassing argument.
If it is copy, it will copy the dispatcher.
If it is new, a new one will be created.
If it is keep, nothing will be done.

The new dispatch methods has now been added to the Atom
class and the Lattice class.

Atom.to also works for changing to a Sphere.

Lattice now supports more transfers.

TODO find a way to document the `to|new` dispatchers.

Fixed imports.

Signed-off-by: Nick Papior <nickpapior@gmail.com>