Added Info attributes to ascii class files
Now all ascii files will contain the info attributes.

This enables the class to dynamically add attributes
based on the content of Sile._info_attributes_.

Each of those entries will be converted to a regex Pattern
and will process on each read line by patching the instance
readline function.

This is very crude, and might be slow for very big files with lots
of attributes, perhaps it can be streamlined a bit by removing those
attributes which are static and already found (or perhaps all non-updatable
attributes will be removed for searching once the full file has been processed).

It proves quite simple to use and makes the code look prettier and
more readable. But the internals are a bit messy.

Signed-off-by: Nick Papior <nickpapior@gmail.com>