fixed #633 due to weird combination of boundary lookups
The problem of close arose when the searched sphere
was just crossing a border of the unit-cell (or supercell).
In those cases could the translated to unit-cell indices
miss a neighbour.

The fix was rather simple; once the atoms closests
has been found, only loop over those in the primary unit-cell.

There are many other small changes here, but they
are more cosmetic and should provide more stability
when using orbital distances very close to atomic coordinates.

Signed-off-by: Nick Papior <nickpapior@gmail.com>