updated a few of the R distances, and added warning when used incorrectly
When using R distances larger than the internal ranges
then the double sphere used in the iter_block_rand may be wrong.
This turns out to be important as some connections won't be reached.
Currently, this is fixed for direct construct calls, but not for
custom self made ones.

As long as the user uses a suitable R range, then it should
work fine.

Signed-off-by: Nick Papior <nickpapior@gmail.com>