Optionally avoid custom node errors.
Custom node errors, when shown directly, obscure the
real cause of the error by only showing it in the handled
exception's traceback. By default, don't raise the custom error
but the original one.

However, still allow the user to request the custom error to
be raised, because it contains useful information like the node
that errored and which inputs was using.