Added parsing of conditional expressions in workflows.
Things like "a = function_1() if test() else function_2()" are now valid
as workflow code. This removes the need for having to call/define a function
that just does the branching.

Furthermore, since the conditional expression is now understood and parsed
into a custom "ConditionalExpressionSyntaxNode", only the branch needed (as
decided by `test()`) is executed.

Comparison syntax (e.g. "a == b") is also parsed into a custom node, therefore
it is now allowed in the syntax of workflows.