added errors explaining which fdf-inputs are required for tbtrans
This should make debugging errors for users simpler in cases
where there is a keyerror, and they think they can just extract
the orbital-transmissions from calculations which didn't calculate it.

Also fixed projections orbital-transmissions. I think the old one
didn't parse the projections correctly, now it should.

warnings can also be issued in case the users
wants to know what is missing etc.

This is still work in progress, but considering we could
add info stuff for the fdf sile.

Signed-off-by: Nick Papior <nickpapior@gmail.com>