major refactor, added sisl._core
Now all sisl core functionality should be
placed in the _core module.
This should adapt the needs for future additions
more easily and aid in the modularization of the code structure.

As it is, nothing needs to be changed in terms of usage.
One thing that does make it simpler to use is that
the aim of sisl is to make the methods that expose a functional
approach be a dispatch method.

For instance:

def tile(geometry: Geometry, ...)
def tile(lattice: Lattice, ...)

are now defined once, and then attached to the
appropriate class of the same name.
Simultaneously the method is hosted in a dispatch
form so that: sisl.tile(geom, ...) can be done.
A call sisl.tile(0, ...) will result in a SislError
since the first argument has not been registered.

As such, one cannot really see anything.

To gain access to the direct method (with appropriate
documentation), simply do:

  geometry_tile = sisl.tile.dispatch(sisl.Geometry)

Signed-off-by: Nick Papior <nickpapior@gmail.com>