added many more functions, also from Grid
These accommodations required an upgrade of
the typing system.

Sadly there is a *bug*/*feature-request*
that broke >=3.9 with missing names.
It is related to this:
 https://bugs.python.org/issue41987

Basically the AtomsArgument can't be used
in the type-hinting because the ForwardRef lookup
tries to resolve it, regardless of whether we are
in a TYPE_CHECKING state or not. Hence it *must*
be defined before we can use it.

This happens, even if __future__ annotations
has been imported.

This forces us to define them via strings manually,
*sigh*.

Signed-off-by: Nick Papior <nickpapior@gmail.com>