trying to fix circular imports in typing constructs
Sadly, the register function of singledispatch does
not handle the ForwardRefs very elegantly
as the objects references *needs* to be available
in the global space.
This is rather unfortunate as this is exactly what
we are trying to avoid using __future__ annotations.

I'll have to see later if this causes issues in the
documentation build.

Signed-off-by: Nick Papior <nickpapior@gmail.com>