fixed memory problems in the finder
In some cases (when the bin-size is big) the
maximum pairs it would allow could be quite big.
When the bins has many atoms, the error in the
actual neighbors can be significant, leading
to memory problems.

Now the finder uses a maximum of 200MB of initial
neighbor lists. This results in incremental re-allocations
which is a bit more heavy, but it reduces the chances of
running out of memory.

One can change the default memory by changing the
class variable: Neighbor.memory accordingly.

Added it to the documentation.

Signed-off-by: Nick Papior <nickpapior@gmail.com>