added typing in lots of io places, added SileSlicer
Added lots of typehints in the code, primarily in the
io code base.

Changed the way the siesta.stdout siles handle data.
They now rely on the SileSlicer and thus can be sliced
upon calling. It makes the code a little simpler but also
highligted some problems.

In particular the problem arise when the return function
needs to do different things depending on whether the file
is done reading, or not.

The defaults for stdoutSileSiesta has changed to read
the next entry in the file. This is in contrast to the
way it was. It will require users to adapt!
the reason for this can be found in the discussion in
#586.

This streamlines the usage across the different parts of the
IO handling.

Signed-off-by: Nick Papior <nickpapior@gmail.com>