fixed #700 by reverting 7cf88f1a5
In 7cf88f1a5243541d9d86fae42c34b518a0cb02ed the introduction
of ufunc.resolve_types was done.
However, this is not a viable approach when dealing with
older numpy versions.
Instead we now fully rely on the return value of the ufunc.
This should bypass any dtype argument handling and hopefully
work with many versions of numpy.

Signed-off-by: Nick Papior <nickpapior@gmail.com>