lots of small updates/fixes
Fixed #708, in cases where the argument was a string
and wasn't part of the Z-tables, then it saved a str
as the Z value. It should *always* be an integer.
Anything beyond the periodic table should be considered
AtomUnknown/AtomGhost.

Removed deprecated set_sc method in LatticeChild (due for 0.14).

Fixed parsing HSX files due to it handling the Geometry.atoms
wrongly, this was partly due to #708, but also a bug had
sneaked in.

Signed-off-by: Nick Papior <nickpapior@gmail.com>